Redesign ExternalName
Allow ExternalName to hold an unrestricted byte array. This is convenient
for numerous use cases, such as using it to hold a symbol name.
Cranelift codegen itself still doesn't interpret these names.

And, adjust the syntax for ExternalNames to account for this. The new
syntax looks like this:

 - `@unquoted_name` (name)
 - `@"quoted name with spaces and other things!"` (name)
 - `@[456:789]` (two-level index)
 - `@<FloorF32>` (libcall)
 - `@-` (default)

This way, all external names start with '@', and we can tell what kind
it is by looking at the following character.

This also changes the source location information prefix to '!'.